fix: fixed all bugs mentioned in the comments
