+++ src/ui/utils/vite-svgr/loadSvgIcon.ts
@@ -1,0 +1,8 @@
+import type { DynamicIcon } from './types/IconType';
+
+export async function loadSvgIcon(iconName: string): Promise<DynamicIcon> {
+  const module = (await import(`../../../assets/icons/${iconName}.svg?react`)) as {
+    default: DynamicIcon;
+  };
+  return module.default;
+}
