+++ src/ui/utils/vite-svgr/types/IconType.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+import type { FunctionComponent } from 'react';
+
+export type DynamicIcon = FunctionComponent<React.SVGProps<SVGSVGElement>> | null;
